refactor: 프로젝트 컨트롤러 폴더 구조 수정
